uapi: change most params type from u32 to u64
Some of these parameters can be set to bigger than max_u32, it makes
sense to change it to u64.

While at it, add validate_address param for future retry use, and
set initial_smoothed_rtt to 333000 by default.

+++ net/quic/stream.h
@@ -43,9 +43,9 @@ struct quic_stream_table {
 	struct quic_hash_table ht;
 
 	struct {
-		u32 max_stream_data_bidi_local;
-		u32 max_stream_data_bidi_remote;
-		u32 max_stream_data_uni;
+		u64 max_stream_data_bidi_local;
+		u64 max_stream_data_bidi_remote;
+		u64 max_stream_data_uni;
 		u64 max_streams_bidi;
 		u64 max_streams_uni;
 		u64 streams_bidi;
@@ -53,9 +53,9 @@ struct quic_stream_table {
 		u64 stream_active;
 	} send;
 	struct {
-		u32 max_stream_data_bidi_local;
-		u32 max_stream_data_bidi_remote;
-		u32 max_stream_data_uni;
+		u64 max_stream_data_bidi_local;
+		u64 max_stream_data_bidi_remote;
+		u64 max_stream_data_uni;
 		u64 max_streams_bidi;
 		u64 max_streams_uni;
 	} recv;
